Clean up project directory and remove unnecessary files
This commit includes the following changes:
- Removed all temporary files and directories that were no longer needed
- Deleted unused libraries and dependencies that were not being used by the project
- Cleared out the pycache directory to remove compiled Python files
- Removed any other files that were not necessary for the project

By removing these unnecessary files, we can improve the overall cleanliness and organization of the project directory. This should make it easier for developers to navigate the codebase and avoid any potential conflicts or confusion.

Overall, this commit should have no impact on the functionality of the project, but it will help to streamline future development efforts and ensure that the project remains easy to work with.